+++ ex05.py
@@ -2,7 +2,7 @@ PARES = []
 IMPARES = []
 print("Escolha 10 números para verificar se são pares ou ímpares: ")
 for i in range(0,10):
-    num = float(input(f"número {i+1}: "))
+    num = int(input(f"número {i+1}: "))
     if num%2 != 0:
         IMPARES.append(num)
     else:
